fix: using sequence table per entity in alovoa_test db
There are some tests that are expecting statics ids for entities, like "UserIntetion" entity. And when it's created with a different id, it throws a Exception and fails the test. So, by using 'standard strategy', it'll create for every entity a sequence table, matching the ids that the test is expecting.

*Note:* this was caused due to upgrade of hibernate, the changes in generations of ids.